+++ app/src/main/kotlin/me/xizzhu/android/joshua/bookmarks/BookmarksListView.kt
@@ -19,25 +19,30 @@ package me.xizzhu.android.joshua.bookmarks
 import android.content.Context
 import android.content.DialogInterface
 import android.util.AttributeSet
-import androidx.recyclerview.widget.LinearLayoutManager
-import androidx.recyclerview.widget.RecyclerView
+import android.view.View
 import me.xizzhu.android.joshua.R
-import me.xizzhu.android.joshua.core.Settings
 import me.xizzhu.android.joshua.core.VerseIndex
 import me.xizzhu.android.joshua.ui.DialogHelper
+import me.xizzhu.android.joshua.ui.recyclerview.BaseRecyclerView
+import me.xizzhu.android.joshua.ui.recyclerview.BookmarkItem
+import me.xizzhu.android.joshua.ui.recyclerview.BookmarkItemViewHolder
 import me.xizzhu.android.joshua.utils.BaseSettingsView
 
 interface BookmarksView : BaseSettingsView {
-    fun onBookmarksLoaded(bookmarks: List<BookmarkForDisplay>)
+    fun onBookmarksLoaded(bookmarks: List<BookmarkItem>)
 
     fun onBookmarksLoadFailed()
 
     fun onVerseSelectionFailed(verseToSelect: VerseIndex)
 }
 
-class BookmarksListView : RecyclerView, BookmarksView {
+class BookmarksListView : BaseRecyclerView, BookmarksView {
     private lateinit var presenter: BookmarksPresenter
-    private val adapter: BookmarksListAdapter
+    private val onClickListener = OnClickListener { view ->
+        ((getChildViewHolder(view) as BookmarkItemViewHolder).item)?.let {
+            presenter.selectVerse(it.verseIndex)
+        }
+    }
 
     constructor(context: Context) : super(context)
 
@@ -45,27 +50,22 @@ class BookmarksListView : RecyclerView, BookmarksView {
 
     constructor(context: Context, attrs: AttributeSet, defStyleAttr: Int) : super(context, attrs, defStyleAttr)
 
-    init {
-        layoutManager = LinearLayoutManager(context, VERTICAL, false)
-
-        adapter = BookmarksListAdapter(context, object : BookmarksListAdapter.Listener {
-            override fun onBookmarkClicked(verseIndex: VerseIndex) {
-                presenter.selectVerse(verseIndex)
-            }
-        })
-        setAdapter(adapter)
-    }
-
     fun setPresenter(presenter: BookmarksPresenter) {
         this.presenter = presenter
     }
 
-    override fun onSettingsUpdated(settings: Settings) {
-        adapter.setSettings(settings)
+    override fun onChildAttachedToWindow(child: View) {
+        super.onChildAttachedToWindow(child)
+        child.setOnClickListener(onClickListener)
     }
 
-    override fun onBookmarksLoaded(bookmarks: List<BookmarkForDisplay>) {
-        adapter.setBookmarks(bookmarks)
+    override fun onChildDetachedFromWindow(child: View) {
+        super.onChildDetachedFromWindow(child)
+        child.setOnClickListener(null)
+    }
+
+    override fun onBookmarksLoaded(bookmarks: List<BookmarkItem>) {
+        setItems(bookmarks)
     }
 
     override fun onBookmarksLoadFailed() {