+++ CHANGELOG.md
@@ -3,6 +3,7 @@ CHANGELOG
 
 #### Next Release
 - Properly catch exception if thrown when downloading translations
+- Refactored to introduce a common adapter
 
 #### v0.1.2 (2019-04-30)
 - Improved verse detail